Plugins::WebBrowser: modify download to replace QWidget with QFrame

+++ Src/ParameterCompone/Stats.cpp
@@ -17,11 +17,11 @@ QString CStats::Convertbytes(quint64 bytes)
     if((1 << 10) >= bytes)
         szBytes = QString::number(bytes) + " " + tr("B");
     else if((1 << 20) >= bytes)
-        szBytes = QString::number(bytes / (1 << 10), 'f', 2) + " " + tr("KB");
+        szBytes = QString::number((qreal)bytes / (1 << 10), 'f', 2) + " " + tr("KB");
     else if((1 << 30) >= bytes)
-        szBytes = QString::number(bytes / (1 << 20), 'f', 2) + " " + tr("MB");
+        szBytes = QString::number((qreal)bytes / (1 << 20), 'f', 2) + " " + tr("MB");
     else
-        szBytes = QString::number(bytes / (1 << 30), 'f', 2) + " " + tr("GB");
+        szBytes = QString::number((qreal)bytes / (1 << 30), 'f', 2) + " " + tr("GB");
     return szBytes;
 }
